Change action return types resolves #3

+++ HotelFinder.WebAPI/Controllers/HotelsController.cs
@@ -10,12 +10,19 @@ using System.Threading.Tasks;
 
 namespace HotelFinder.WebAPI.Controllers
 {
+    /// <summary>
+    /// API for Hotels
+    /// </summary>
     [Route("api/[controller]")]
     [ApiController]
     public class HotelsController : ControllerBase
     {
         private IHotelService _hotelService;
         
+        /// <summary>
+        /// Takes hotels from hotel service.
+        /// </summary>
+        /// <param name="hotelService"></param>
         public HotelsController(IHotelService hotelService)
         {
             _hotelService = hotelService; // dependency injection
@@ -26,52 +33,74 @@ namespace HotelFinder.WebAPI.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpGet]
-        public List<Hotel> Get()
+        public IActionResult Get()
         {
-            return _hotelService.GetAllHotels();
+            var hotels = _hotelService.GetAllHotels();
+            return Ok(hotels); // 200 + data
         }
 
         /// <summary>
-        /// Gets hotel by id.
+        /// Gets the hotel by id.
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
         [HttpGet("{id}")]
-        public Hotel GetById(int id)
+        public IActionResult GetById(int id)
         {
-            return _hotelService.GetHotelById(id);
+            var hotel = _hotelService.GetHotelById(id);
+            if (hotel != null) 
+            {
+                return Ok(hotel); // 200 + data
+            }
+            return NotFound(); // 404
         }
 
         /// <summary>
-        /// Create hotel.
+        /// Creates a hotel.
         /// </summary>
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPost]
-        public Hotel Post([FromBody]Hotel hotel)
+        public IActionResult Post([FromBody]Hotel hotel)
         {
-            return _hotelService.CreateHotel(hotel);
+            // [ApiController] annotation already controls validation
+            // We can remove this control from here
+            //if (ModelState.IsValid)
+            //{
+                var createdHotel = _hotelService.CreateHotel(hotel);
+                return CreatedAtAction("Get", new { id = createdHotel.Id }, createdHotel ); // 201 + data
+            //}
+            //return BadRequest(ModelState); // 400 + validation errors
         }
 
         /// <summary>
-        /// Update the hotel from body.
+        /// Updates the hotel from body.
         /// </summary>
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPut]
-        public Hotel Put([FromBody]Hotel hotel)
+        public IActionResult Put([FromBody]Hotel hotel)
         {
-            return _hotelService.UpdateHotel(hotel);
+            if (_hotelService.GetHotelById(hotel.Id) != null) 
+            {
+                return Ok(_hotelService.UpdateHotel(hotel)); // 200 + data
+            }
+            return NotFound(); // 404
         }
 
         /// <summary>
-        /// Delete the hotel by id.
+        /// Deletes the hotel by id.
         /// </summary>
         /// <param name="id"></param>
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
-            _hotelService.DeleteHotel(id);
+            if (_hotelService.GetHotelById(id) != null)
+            {
+                _hotelService.DeleteHotel(id);
+                return Ok(); // 200
+            }
+            return NotFound(); // 404
         }
     }
 }
